requirements.txt
web3==6.19.0
python-dotenv==1.0.0
requests==2.31.0
loguru==0.7.2
python-telegram-bot==20.7